+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/provider/FeignRequestTemplateMetadataProvider.java
@@ -43,7 +43,7 @@ public class FeignRequestTemplateMetadataProvider implements MetadataProvider {
 	}
 
 	@Override
-	public String getRawMetadataStringValue(String key) {
+	public String doGetRawMetadataStringValue(String key) {
 		switch (key) {
 		case MessageMetadataContainer.LABEL_KEY_METHOD:
 			return requestTemplate.method();
@@ -59,7 +59,7 @@ public class FeignRequestTemplateMetadataProvider implements MetadataProvider {
 	}
 
 	@Override
-	public String getRawMetadataMapValue(String key, String mapKey) {
+	public String doGetRawMetadataMapValue(String key, String mapKey) {
 		Map<String, Collection<String>> headers = requestTemplate.headers();
 		switch (key) {
 		case MessageMetadataContainer.LABEL_MAP_KEY_HEADER: