+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/util/PolarisCompletableFutureUtilsTest.java
@@ -17,13 +17,20 @@
 
 package com.tencent.cloud.common.util;
 
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
+
 import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import org.springframework.util.concurrent.FutureUtils;
+
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * Test for {@link PolarisCompletableFutureUtils}.
@@ -45,6 +52,22 @@ public class PolarisCompletableFutureUtilsTest {
 
 	@Test
 	public void testSupplyAsync() {
+		// can not be found and set
+		CompletableFuture.supplyAsync(() -> {
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isNull();
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+			return "test";
+		}).thenAccept(result -> {
+			assertThat(result).isEqualTo("test");
+		}).join();
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+
+		// can be found and set
 		PolarisCompletableFutureUtils.supplyAsync(() -> {
 			assertThat(MetadataContextHolder.get()
 					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
@@ -62,7 +85,27 @@ public class PolarisCompletableFutureUtilsTest {
 	}
 
 	@Test
+	public void testSupplyAsyncWithNullSupplier() {
+		assertThatThrownBy(() -> PolarisCompletableFutureUtils.supplyAsync(null))
+				.isExactlyInstanceOf(IllegalArgumentException.class)
+				.hasMessageContaining("Supplier must not be null");
+	}
+
+	@Test
 	public void testRunAsync() {
+		// can not be found and set
+		CompletableFuture.runAsync(() -> {
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isNull();
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+		}).join();
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+
+		// can be found and set
 		PolarisCompletableFutureUtils.runAsync(() -> {
 			assertThat(MetadataContextHolder.get()
 					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
@@ -74,5 +117,100 @@ public class PolarisCompletableFutureUtilsTest {
 				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
 		assertThat(MetadataContextHolder.get()
 				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isEqualTo("value3");
+	}
+
+	@Test
+	public void testRunAsyncWithNullRunnable() {
+		assertThatThrownBy(() -> PolarisCompletableFutureUtils.runAsync(null))
+				.isExactlyInstanceOf(IllegalArgumentException.class)
+				.hasMessageContaining("Runnable must not be null");
+	}
+
+	@Test
+	public void testCallAsync() {
+		// can not be found and set
+		FutureUtils.callAsync(() -> {
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isNull();
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+			return "test";
+		}).thenAccept(result -> {
+			assertThat(result).isEqualTo("test");
+		}).join();
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+
+		// can be found and set
+		PolarisCompletableFutureUtils.callAsync(() -> {
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+			MetadataContextHolder.get().putContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2", "value2");
+			return "test";
+		}).thenAccept(result -> {
+			assertThat(result).isEqualTo("test");
+		}).join();
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isEqualTo("value2");
+	}
+
+	@Test
+	public void testCallAsyncWithNullCallable() {
+		assertThatThrownBy(() -> PolarisCompletableFutureUtils.callAsync(null))
+				.isExactlyInstanceOf(IllegalArgumentException.class)
+				.hasMessageContaining("Callable must not be null");
+	}
+
+	@Test
+	public void testCallAsyncWithExecutor() {
+		Executor executor = Executors.newSingleThreadExecutor();
+
+		// can not be found and set
+		FutureUtils.callAsync(() -> {
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isNull();
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+			return "test";
+		}, executor).thenAccept(result -> {
+			assertThat(result).isEqualTo("test");
+		}).join();
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+
+		// can be found and set
+		PolarisCompletableFutureUtils.callAsync(() -> {
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+			assertThat(MetadataContextHolder.get()
+					.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isNull();
+			MetadataContextHolder.get().putContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2", "value2");
+			return "test";
+		}, executor).thenAccept(result -> {
+			assertThat(result).isEqualTo("test");
+		}).join();
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key1")).isEqualTo("value1");
+		assertThat(MetadataContextHolder.get()
+				.getContext(MetadataContext.FRAGMENT_TRANSITIVE, "key2")).isEqualTo("value2");
+	}
+
+	@Test
+	public void testCallAsyncWithNullExecutor() {
+		assertThatThrownBy(() -> PolarisCompletableFutureUtils.callAsync(null, null))
+				.isExactlyInstanceOf(IllegalArgumentException.class)
+				.hasMessageContaining("Callable must not be null");
+
+		assertThatThrownBy(() -> PolarisCompletableFutureUtils.callAsync(() -> "test", null))
+				.isExactlyInstanceOf(IllegalArgumentException.class)
+				.hasMessageContaining("Executor must not be null");
 	}
 }