+++ CHANGELOG.md
@@ -39,3 +39,4 @@
 - [feat:support body store switch.](https://github.com/Tencent/spring-cloud-tencent/pull/1771)
 - [refactor:optimize auto configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1739)
 - [refactor:optimize config locate.](https://github.com/Tencent/spring-cloud-tencent/pull/1741)
+- [feat:support async metadata transfer.](https://github.com/Tencent/spring-cloud-tencent/pull/1743)